requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
